♻️refactor: Print 부분 분리 & view로 이동
