+++ .idea/workspace.xml
@@ -6,6 +6,8 @@
   <component name="ChangeListManager">
     <list default="true" id="37600dd4-43dd-4b2d-a62d-19fc13fe37ec" name="변경" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/LadderGame/Main.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/LadderGame/Main.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/LadderGame/domain/Ladder.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/LadderGame/domain/Ladder.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
