+++ src/main/java/LadderGame/view/Print.java
@@ -1,0 +1,40 @@
+package LadderGame.view;
+
+import LadderGame.domain.Ladder;
+import LadderGame.domain.Line;
+
+import java.util.ArrayList;
+
+public class Print {
+    // 이름 출력
+    public void namePrint(ArrayList<String> names) {
+        for (String temp : names) {
+            System.out.print(temp);
+            spacePrint(temp.length());
+        }
+        System.out.println();
+    }
+
+    // 빈 공간 출력
+    public void spacePrint(int length) {
+        for (int i = 6 - length; i > 0; i--) {
+            System.out.print(" ");
+        }
+    }
+
+    // Boolean(true) -> String("-----")
+    public String boolToString(boolean point) {
+        if (point) return "-----";
+        return "     ";
+    }
+
+    // 사다리 출력
+    public void setPrint(Ladder ladder) {
+        namePrint(ladder.getNames());
+
+        for (Line line : ladder.getLadder_set()) {
+            line.getPoints().forEach(point -> System.out.print("|" + boolToString(point)));
+            System.out.println("|");
+        }
+    }
+}
